feat: rename project to "mlrisko" and update related configurations
- Renamed the project from "risk_control" to "mlrisko" across various files including `.gitignore`, `Makefile`, `pyproject.toml`, and `README.md`.
- Updated CI configuration to reflect the new project name.
- Adjusted import statements in example scripts and modules to align with the new naming convention.
- Enhanced project metadata in `pyproject.toml` and updated the README to describe the toolkit's purpose more clearly.
