+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "riskcontrol-todo"
+name = "mlrisko"
 version = "0.1.0"
-description = "Calibrate predictive algorithms to achieve risk control."
+description = "A toolkit to calibrate predictive algorithms to achieve risk control."
 keywords = ["risk-control", "machine-learning", "ai", "statistics", "multiple-testing"]
 readme = "README.md"
 license = "BSD-3-Clause"
@@ -20,7 +20,6 @@ dependencies = [
 ]
 classifiers = [
     "Programming Language :: Python :: 3.13",
-    "License :: OSI Approved :: BSD License",
     "Operating System :: OS Independent",
     "Development Status :: 2 - Pre-Alpha",
     "Intended Audience :: Science/Research",
@@ -34,7 +33,7 @@ classifiers = [
 "Source Code" = "https://github.com/thibaultcordier/risk-control"
 "Documentation" = "https://thibaultcordier.github.io/risk-control/"
 "Changelog" = "https://github.com/thibaultcordier/risk-control/releases"
-"PyPI" = "https://pypi.org/project/riskcontrol-todo/"
+"PyPI" = "https://pypi.org/project/mlrisko/"
 "Code Coverage" = "https://codecov.io/gh/thibaultcordier/risk-control"
 "CI/CD Pipeline" = "https://github.com/thibaultcordier/risk-control/actions"
 
@@ -57,8 +56,15 @@ dev = [
     "scipy-stubs>=1.16.0.2",
 ]
 
+[build-system]
+requires = ["uv_build >= 0.7.19, <0.9.0"]
+build-backend = "uv_build"
+
 [tool.uv]
 package = false
+
+[tool.uv.build-backend]
+module-root = ""
 
 [tool.mypy]
 python_version = "3.13"
@@ -70,7 +76,7 @@ disallow_untyped_defs = true
 addopts = "-vs -vv -vsx"
 testpaths = ["tests"]
 python_files = ["test_*.py"]
-pythonpath = "risk_control"
+pythonpath = "mlrisko"
 
 [tool.ruff]
 line-length = 88
