+++ tests/test_imports.py
@@ -7,35 +7,35 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 
 class TestModuleImports(unittest.TestCase):
-    """Test cases for importing risk_control modules."""
+    """Test cases for importing mlrisko modules."""
 
     def test_parameter_module(self) -> None:
         """Test the parameter module directly."""
         # Let's check what's available in parameter module
-        import risk_control.parameter
+        import mlrisko.parameter
 
-        self.assertIsNotNone(risk_control.parameter)
+        self.assertIsNotNone(mlrisko.parameter)
 
     def test_abstention_module(self) -> None:
         """Test the abstention module directly."""
         # Let's check what's available in abstention module
-        import risk_control.abstention
+        import mlrisko.abstention
 
-        self.assertIsNotNone(risk_control.abstention)
+        self.assertIsNotNone(mlrisko.abstention)
 
     def test_plot_module(self) -> None:
         """Test the plot module directly."""
         # Let's check what's available in plot module
-        import risk_control.plot
+        import mlrisko.plot
 
-        self.assertIsNotNone(risk_control.plot)
+        self.assertIsNotNone(mlrisko.plot)
 
     def test_risk_module(self) -> None:
         """Test the risk module directly."""
         # Let's check what's available in risk module
-        import risk_control.risk
+        import mlrisko.risk
 
-        self.assertIsNotNone(risk_control.risk)
+        self.assertIsNotNone(mlrisko.risk)
 
 
 if __name__ == "__main__":
